+++ src/controllers/auth.js
@@ -12,6 +12,7 @@ const UserVote = mongoose.model('UserVote');
 const App = mongoose.model('App');
 const UserGoal = mongoose.model('UserGoal');
 const UserBookRating = mongoose.model('UserBookRating');
+const UserComment = mongoose.model('UserComment');
 
 const checkAuth = async (req, res) => {
   const { token } = req.query;
@@ -28,15 +29,17 @@ const checkAuth = async (req, res) => {
           User.findOneAndUpdate({ _id: userId }, { lastLoggedIn }),
           UserVote.find({ userId }).select({ bookId: 1, count: 1 }),
           UserGoal.find({ userId }).select({ numberOfPages: 1 }),
-          UserBookRating.find({ userId }).select({ bookId: 1, rating: 1 })
+          UserBookRating.find({ userId }).select({ bookId: 1, rating: 1 }),
+          UserComment.find({ userId }).select({ bookId: 1, added: 1, comment: 1 })
         ]);
         const { version, googlePlayUrl } = result[0][0] || {};
         const { _id, email, registered, updated, supportApp } = result[1] || {};
         const userVotes = result[2] || 0;
         const { numberOfPages } = result[3][0] || 0;
         const userBookRatings = result[4] || 0;
+        const userComments = result[5] || 0;
 
-        res.send({ profile: { _id, email, registered, updated, supportApp }, version, googlePlayUrl, userVotes, userBookRatings, numberOfPagesForGoal: numberOfPages });
+        res.send({ profile: { _id, email, registered, updated, supportApp }, version, googlePlayUrl, userVotes, userBookRatings, numberOfPagesForGoal: numberOfPages, userComments });
       } catch (err) {
         return res.status(500).send({
           fieldName: 'other',
@@ -128,13 +131,19 @@ const signIn = async (req, res) => {
       const token = jwt.sign({ userId }, 'I_LIKE_READING_BOOKS_209');
       const lastLoggedIn = currentDate.getTime();
       await User.findOneAndUpdate({ _id: userId }, { lastLoggedIn });
-      const userVotes = await UserVote.find({ userId }).select({ bookId: 1, count: 1 });
-      const appInfo = await App.find({}).select({ version: 1, googlePlayUrl: 1 });
-      const userGoal = await UserGoal.find({ userId }).select({ numberOfPages: 1 });
-      const userBookRatings = await UserBookRating.find({ userId }).select({ bookId: 1, rating: 1 });
-      const { version, googlePlayUrl } = appInfo[0] || {};
-      const { numberOfPages } = userGoal[0] || {};
-      return res.send({ token, profile, version, googlePlayUrl, userVotes, userBookRatings, numberOfPagesForGoal: numberOfPages });
+      const result = await Promise.all([
+        App.find({}).select({ version: 1, googlePlayUrl: 1 }),
+        UserVote.find({ userId }).select({ bookId: 1, count: 1 }),
+        UserGoal.find({ userId }).select({ numberOfPages: 1 }),
+        UserBookRating.find({ userId }).select({ bookId: 1, rating: 1 }),
+        UserComment.find({ userId }).select({ bookId: 1, added: 1, comment: 1 })
+      ]);
+      const { version, googlePlayUrl } = result[0][0] || {};
+      const userVotes = result[1] || 0;
+      const { numberOfPages } = result[2][0] || 0;
+      const userBookRatings = result[3] || 0;
+      const userComments = result[4] || 0;
+      return res.send({ token, profile, version, googlePlayUrl, userVotes, userBookRatings, numberOfPagesForGoal: numberOfPages, userComments });
     } catch (e) {
       return res.status(500).send({
         fieldName: 'other',
@@ -178,18 +187,25 @@ const signIn = async (req, res) => {
       });
     }
     await user.comparePassword(password);
-    const userVotes = await UserVote.find({ userId: user._id }).select({ bookId: 1, count: 1 });
-    const userBookRatings = await UserBookRating.find({ userId: user._id }).select({ bookId: 1, rating: 1 });
-    const token = jwt.sign({ userId: user._id }, 'I_LIKE_READING_BOOKS_209');
+    
     profile = { _id: user._id, email: user.email, registered: user.registered, updated: user.updated, supportApp: user.supportApp };
     const currentDate = new Date();
     const lastLoggedIn = currentDate.getTime();
     await User.findOneAndUpdate({ _id: user._id }, { lastLoggedIn });
-    const appInfo = await App.find({}).select({ version: 1, googlePlayUrl: 1 });
-    const userGoal = await UserGoal.find({ userId: user._id }).select({ numberOfPages: 1 });
-    const { version, googlePlayUrl } = appInfo[0] || {};
-    const { numberOfPages } = userGoal[0] || {};
-    return res.send({ token, profile, userVotes, userBookRatings, version, googlePlayUrl, numberOfPagesForGoal: numberOfPages });
+    const result = await Promise.all([
+      App.find({}).select({ version: 1, googlePlayUrl: 1 }),
+      UserVote.find({ userId: user._id }).select({ bookId: 1, count: 1 }),
+      UserGoal.find({ userId: user._id }).select({ numberOfPages: 1 }),
+      UserBookRating.find({ userId: user._id }).select({ bookId: 1, rating: 1 }),
+      UserComment.find({ userId: user._id }).select({ bookId: 1, added: 1, comment: 1 })
+    ]);
+    const { version, googlePlayUrl } = result[0][0] || {};
+    const userVotes = result[1] || 0;
+    const { numberOfPages } = result[2][0] || 0;
+    const userBookRatings = result[3] || 0;
+    const userComments = result[4] || 0;
+    const token = jwt.sign({ userId: user._id }, 'I_LIKE_READING_BOOKS_209');
+    return res.send({ token, profile, userVotes, userBookRatings, version, googlePlayUrl, numberOfPagesForGoal: numberOfPages, userComments });
   } catch (err) {
     console.log(err, 'err');
     return res.status(500).send({